PDFBOX-3082: reduce memory on IndexData, as suggested by Tom Callahan
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1714941 13f79535-47bb-0310-9956-ffa450edef68

+++ fontbox/src/main/java/org/apache/fontbox/cff/CFFParser.java
@@ -163,7 +163,7 @@ public class CFFParser
         index.initData(dataSize);
         for (int i = 0; i < dataSize; i++)
         {
-            index.setData(i, input.readCard8());
+            index.setData(i, (byte) input.readCard8());
         }
         return index;
     }
